solucionar errores
