--- app.py
@@ -1,9 +1,0 @@
-
-from crear_base_usuarios import contrasena_crear;
-
-
-from inicioSesion import inicio_sesion;
-
-contrasena_crear()
-
-inicio_sesion()
